requirements.txt
fastapi
uvicorn
meilisearch
sentence-transformers
faiss-cpu
pandas
python-dotenv
